+++ resources/lang/en/validation/endBorrowing.php
@@ -2,5 +2,5 @@
 
 return [
     'selectedBorrowings.required' => 'Select borrowings to end.',
-    'selectedBorrowings.*.distinct' => 'The borrowing :value was filled in multiple times.' // TODO : Create custom validator to feed the name instead of the id to the message
+    'selectedBorrowings.*.distinct' => 'The borrowing of :item by :borrower was selected in multiple times.'
 ];