+++ resources/lang/fr/validation/endBorrowing.php
@@ -2,5 +2,5 @@
 
 return [
     'selectedBorrowings.required' => 'Sélectionnez des emprunts à terminer.',
-    'selectedBorrowings.*.distinct' => 'L\'emprunt du jeu :value a été renseigné plusieurs fois.' // TODO : Create custom validator to feed the name instead of the id to the message
+    'selectedBorrowings.*.distinct' => 'L\'emprunt du jeu :item par :borrower a été choisi plusieurs fois.'
 ];